pyproject.toml
[tool.poetry]
name = "site-flask"
version = "0.1.0"
description = ""
authors = ["asFrants <as.fratnsev@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Flask = "2.0.2"
Werkzeug = "2.0.3"
Jinja2 = "3.0.2"
Flask-Login = "0.5.0"
Flask-Migrate = "3.1.0"
Flask-WTF = "1.0.0"
Flask-SQLAlchemy = "2.5.1"
SQLAlchemy = "1.4.29"
email-validator = "1.1.3"
gunicorn = "20.1.0"
flask-restx = "0.5.1"
python-dotenv = "0.19.2"
Flask-Minify = "0.37"
Flask-Bootstrap = "^3.3.7"
Flask-Moment = "^1.0.5"
WTForms = "^3.0.1"
alembic = "^1.9.2"
blinker = "^1.5"
requests = "^2.28.2"
telebot = "^0.0.5"
loguru = "^0.7.0"
wheel = "^0.40.0"
pg8000 = "^1.29.4"
bcrypt = "^4.0.1"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

#[build-system]
#build-backend = 'setuptools.build_meta'
#requires = ['setuptools >= 65.5.0',]